// src/types/flags.ts
export interface GlobalFlagsInterface {
  tfInnerBatchTxn?: boolean;
}
// ...
export interface AMMWithdrawFlagsInterface extends GlobalFlagsInterface {
  tfLPToken?: boolean;
  tfWithdrawAll?: boolean;
  tfOneAssetWithdrawAll?: boolean;
  tfSingleAsset?: boolean;
  tfTwoAsset?: boolean;
  tfOneAssetLPToken?: boolean;
  tfLimitLPToken?: boolean;
}